sigh -- keep forgetting refactor stuff
