+++ app/src/main/java/com/aa/draftt/views/onboarding/OnboardingScreenFragment.kt
@@ -1,4 +1,4 @@
-package com.aa.draftt.onboarding.views
+package com.aa.draftt.views.onboarding
 
 import android.content.Intent
 import android.os.Bundle
